Tests for delete user API

+++ src/main/java/gugunava/danil/usermanagementservice/service/UserService.java
@@ -75,7 +75,8 @@ public class UserService {
     }
 
     public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+        if (userRepository.existsById(id))
+            userRepository.deleteById(id);
     }
 
     private boolean isCommandEmpty(UpdateUserCommand command) {
